pyproject.toml
[project]
name = "neaurlscript-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.1",
    "fastapi[standard]>=0.116.1",
    "itsdangerous>=2.2.0",
    "python-dotenv>=1.1.1",
]
